+++ src/service-implementations/lf-fields.service.ts
@@ -71,17 +71,27 @@ export class LfFieldsService implements LfFieldContainerService {
   }
 
   async getUnfilteredFieldDefinitionsAsync(): Promise<FieldDefinition[]> {
-    if (this.cachedFieldDefinitions) {
-      return this.cachedFieldDefinitions;
-    } else {
+    if (!this.cachedFieldDefinitions) {
+      const resultFieldDefinitions: FieldDefinition[] = [];
+      const callback = async (response: FieldDefinitionCollectionResponse): Promise<boolean> => {
+        if (!response.value || response.value.length === 0) {
+          return false;
+        }
+
+        resultFieldDefinitions.push(...response.value);
+        return true;
+      };
+
       const repositoryId = await this.repoClient.getCurrentRepoId();
-      const response: FieldDefinitionCollectionResponse =
-        await this.repoClient.fieldDefinitionsClient.listFieldDefinitions({
-          repositoryId,
-        });
-      this.cachedFieldDefinitions = response.value ?? [];
-      return response.value ?? [];
+      await this.repoClient.fieldDefinitionsClient.listFieldDefinitionsForEach({
+        callback,
+        repositoryId,
+        maxPageSize: this.defaultPageSize,
+      });
+
+      this.cachedFieldDefinitions = resultFieldDefinitions;
     }
+    return this.cachedFieldDefinitions;
   }
 
   async getDefaultFieldValuesAsync(): Promise<FieldValue[]> {
@@ -92,13 +102,13 @@ export class LfFieldsService implements LfFieldContainerService {
 
   async getAvailableTemplatesAsync(): Promise<TemplateInfo[]> {
     if (!this.cachedTemplateDefinitions) {
-      let resultTemplateDefinitions: TemplateDefinition[] = [];
+      const resultTemplateDefinitions: TemplateDefinition[] = [];
       const callback = async (response: TemplateDefinitionCollectionResponse): Promise<boolean> => {
         if (!response.value || response.value.length === 0) {
           return false;
         }
 
-        resultTemplateDefinitions = [...resultTemplateDefinitions, ...response.value];
+        resultTemplateDefinitions.push(...response.value);
         return true;
       };